Update parsePEASS.sh
Small redesign and fixups.  See changelog.

+++ parsePEASS/parsePEASS.sh
@@ -2,7 +2,7 @@
 
 # -------------------------------------------------------------------
 # sh = PEASS_ng Output file Parser and Converter
-# v = 1.0
+# v = 1.1
 # id = malloc84 || mnemonic-re @ github
 # -------------------------------------------------------------------
 
@@ -18,15 +18,13 @@ BLUE='\033[0;34m'
 echo "$title"
 PS3="$prompt "
 
-echo -e "Convert PEASS-ng (winpeas or linpeas) output file to HTML or PDF readable format"
-echo ''
-echo -e "${RED}DevNote: PDF conversion might have some issues with python reportlab library.${NO}"
-echo ''
+echo -e "Convert PEASS_ng (linpeas or winpeas) output file to HTML or PDF readable format"
 echo -e "Select the ${YELLOW}*.out ${NO}PEASS output file to process."
 echo ''
 
 files=$(ls *.out)
 i=1
+echo "Found outputs:"
 for j in $files
 do
 echo "$i.$j"
@@ -40,9 +38,9 @@ echo -e "File to process: "${YELLOW}"${file[$input]}"${NO}""
 
 echo ''
 echo -e "${BLUE}Output file conversion options: ${NO}"
-echo -e "${GREEN}1 - Convert ${YELLOW}"${file[$input]}" ${GREEN}file to readable HTML file."
-echo -e "${GREEN}2 - Convert ${YELLOW}"${file[$input]}" ${GREEN}file to readable PDF file."
-echo -e "${GREEN}3 - Convert ${YELLOW}"${file[$input]}" ${GREEN}file to parsable JSON file."
+echo -e "${GREEN}1 - Convert ${YELLOW}"${file[$input]}" ${GREEN}to readable HTML file."
+echo -e "${GREEN}2 - Convert ${YELLOW}"${file[$input]}" ${GREEN}to readable PDF file."
+echo -e "${GREEN}3 - Convert ${YELLOW}"${file[$input]}" ${GREEN}to parsable JSON file."
 echo -e "${BLUE}4 - Restart Conversion script to process another file.${NO}"
 echo -e "${RED}5 - Exit Conversion script.${NO}"
 echo ''
@@ -53,40 +51,59 @@ do
     case $opt in
         "HTML")
             echo -e "${GREEN}[+] Converting RAW output to readable HTML..."
+            if [[ -f peass.json ]]
+            then
+                rm peass.json
+            fi
             sleep 1
             echo -e "${GREEN}[+] RAW output to JSON..."
             python3 parsers/peas2json.py ${file[$input]} peass.json
-            sleep 5
+            if [[ -f peass.json ]]
+            then
+                echo -e "${GREEN}[+] Searching and Fixing BAD characters..."
+                sed -i 's/S<s/Ss/g' peass.json
+                sleep 3
+            fi
             echo -e "${GREEN}[+] JSON to HTML..."
             python3 parsers/json2html.py peass.json ${file[$input]}.html
-            sleep 5
+            sleep 3
             if [[ ! -f ${file[$input]}.html ]]
             then
                 echo -e "${RED}[-] Failed to convert to HTML!${NO}"
             fi
-            rm peass.json
+
             if [[ -f ${file[$input]}.html ]]
             then
+                rm peass.json
                 echo -e "${GREEN}[+] Conversion to ${BLUE}HTML ${GREEN}Finished.${NO}"
                 echo -e "${GREEN}[+] Created:"${YELLOW} ${file[$input]}.html"${NO}"
             fi
             ;;
         "PDF")
             echo -e "${GREEN}[+] Converting RAW output to readable PDF..."
+            if [[ -f peass.json ]]
+            then
+                rm peass.json
+            fi
             sleep 1
             echo -e "${GREEN}[+] RAW output to JSON..."
             python3 parsers/peas2json.py ${file[$input]} peass.json
-            sleep 5
-            echo -e "${GREEN}[+] JSON to PDF..."
+            if [[ -f peass.json ]]
+            then
+                echo -e "${GREEN}[+] Searching and Fixing BAD characters..."
+                sed -i 's/S<s/Ss/g' peass.json
+                sleep 3
+            fi
+            echo -e "${GREEN}[+] JSON to PDF... (might take a while)"
             python3 parsers/json2pdf.py peass.json ${file[$input]}.pdf
-            sleep 5
+            sleep 3
             if [[ ! -f ${file[$input]}.pdf ]]
             then
                 echo -e "${RED}[-] Failed to convert to PDF!${NO}"
             fi
-            rm peass.json
             if [[ -f ${file[$input]}.pdf ]]
             then
+                rm peass.json
                 echo -e "${GREEN}[+] Conversion to ${BLUE}PDF ${GREEN}Finished.${NO}"
                 echo -e "${GREEN}[+] Created:"${YELLOW} ${file[$input]}.pdf"${NO}"
             fi
@@ -94,8 +111,18 @@ do
         "JSON")
             echo -e "${GREEN}[+] Converting RAW output to parsable JSON..."
             sleep 1
+            if [[ -f peass.json ]]
+            then
+                rm peass.json
+            fi
             echo -e "${GREEN}[+] RAW output to JSON..."
             python3 parsers/peas2json.py ${file[$input]} peass.json
+            if [[ -f peass.json ]]
+            then
+                echo -e "${GREEN}[+] Searching and Fixing BAD characters..."
+                sed -i 's/S<s/Ss/g' peass.json
+                sleep 3
+            fi
             sleep 3
             if [[ ! -f peass.json ]]
             then
@@ -104,7 +131,7 @@ do
             if [[ -f peass.json ]]
             then
                 echo -e "${GREEN}[+] Conversion to ${BLUE}JSON ${GREEN}Finished.${NO}" 
-                echo -e "${GREEN}[+] Created:"${YELLOW} ${file[$input]}.json"${NO}"
+                echo -e "${GREEN}[+] Created: "${YELLOW}'peass.json'"${NO}"
             fi
             ;;
         "RESTART")
